+++ index.html
@@ -75,7 +75,7 @@
       </footer>
     </main>
     <script type="module" src="static/js/index.js"></script>
-    <script src="static/js/nav.js"></script>
+    <script type="module" src="static/js/nav.js"></script>
 
 
   </body>
