--- modules/luxon/README.md
@@ -1,55 +1,0 @@
-# Luxon
-
-[![MIT License][license-image]][license] [![Build Status][github-action-image]][github-action-url] [![NPM version][npm-version-image]][npm-url] [![Coverage Status][test-coverage-image]][test-coverage-url] [![PRs welcome][contributing-image]][contributing-url]
-
-Luxon is a library for working with dates and times in JavaScript.
-
-```js
-DateTime.now().setZone("America/New_York").minus({ weeks: 1 }).endOf("day").toISO();
-```
-
-## Upgrading to 2.0
-
-[Guide](https://moment.github.io/luxon/#upgrading)
-
-## Features
- * DateTime, Duration, and Interval types.
- * Immutable, chainable, unambiguous API.
- * Parsing and formatting for common and custom formats.
- * Native time zone and Intl support (no locale or tz files).
-
-## Download/install
-
-[Download/install instructions](https://moment.github.io/luxon/#/install)
-
-## Documentation
-
-* [General documentation](https://moment.github.io/luxon/#/?id=luxon)
-* [API docs](https://moment.github.io/luxon/api-docs/index.html)
-* [Quick tour](https://moment.github.io/luxon/#/tour)
-* [For Moment users](https://moment.github.io/luxon/#/moment)
-* [Why does Luxon exist?](https://moment.github.io/luxon/#/why)
-* [A quick demo](https://moment.github.io/luxon/demo/global.html)
-
-## Development
-
-See [contributing](CONTRIBUTING.md).
-
-![Phasers to stun][phasers-image]
-
-[license-image]: http://img.shields.io/badge/license-MIT-blue.svg
-[license]: LICENSE.md
-
-[github-action-image]: https://github.com/moment/luxon/actions/workflows/test.yml/badge.svg
-[github-action-url]: https://github.com/moment/luxon/actions/workflows/test.yml
-
-[npm-url]: https://npmjs.org/package/luxon
-[npm-version-image]: https://badge.fury.io/js/luxon.svg
-
-[test-coverage-url]: https://codecov.io/gh/moment/luxon
-[test-coverage-image]: https://codecov.io/gh/moment/luxon/branch/master/graph/badge.svg
-
-[contributing-url]: https://github.com/moment/luxon/blob/master/contributing.md
-[contributing-image]: https://img.shields.io/badge/PRs-welcome-brightgreen.svg
-
-[phasers-image]: https://img.shields.io/badge/phasers-stun-brightgreen.svg
