--- modules/luxon/src/impl/invalid.js
@@ -1,13 +1,0 @@
-export default class Invalid {
-  constructor(reason, explanation) {
-    this.reason = reason;
-    this.explanation = explanation;
-  }
-
-  toMessage() {
-    if (this.explanation) {
-      return `${this.reason}: ${this.explanation}`;
-    }
-    return this.reason;
-  }
-}
