+++ static/css/index.css
@@ -1,90 +1,90 @@
 * {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-  }
-  
-  body {
-    display: flex;
-    justify-content: center;
-    padding: 9px;
-  }
-  
-  .wrapper {
-    width: 97%;
-  }
-  
-  .books {
-    border: 3px solid #000;
-    list-style-type: none;
-    padding: 7px;
-  }
-  
-  #form {
-    text-align: center;
-    width: 50%;
-    margin: 0 auto;
-    display: none;
-  }
-  
-  .form-heading {
-    margin: 10px;
-  }
-  
-  input {
-    width: 100%;
-    padding: 5px;
-    border: 2px solid #000;
-  }
-  
-  .button-container {
-    text-align: end;
-  }
-  
-  #addbtn {
-    width: 25%;
-    border: 2px solid #000;
-    padding: 5px;
-    box-shadow: 2px 1px 8px #000;
-  }
-  
-  .remove {
-    padding: 3px;
-    margin: 5px;
-    box-shadow: 2px 1px 8px #000;
-    border: 2px solid #000;
-  }
-  
-  .card {
-    background-color: #fff;
-    display: flex;
-  }
-  
-  .btn {
-    flex: 1;
-    text-align: end;
-  }
-  
-  .p-content {
-    vertical-align: middle;
-  }
-  
-  #message {
-    height: 10px;
-    display: none;
-    color: red;
-    font-size: small;
-    margin: 10px 0 10px;
-  }
-  
-  .container {
-    display: block;
-  }
-  
-  .contact-info-container {
-    display: none;
-  }
-  
-  .add-book {
-    display: none;
-  }
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  display: flex;
+  justify-content: center;
+  padding: 9px;
+}
+
+.wrapper {
+  width: 97%;
+}
+
+.books {
+  border: 3px solid #000;
+  list-style-type: none;
+  padding: 7px;
+}
+
+#form {
+  text-align: center;
+  width: 50%;
+  margin: 0 auto;
+  display: none;
+}
+
+.form-heading {
+  margin: 10px;
+}
+
+input {
+  width: 100%;
+  padding: 5px;
+  border: 2px solid #000;
+}
+
+.button-container {
+  text-align: end;
+}
+
+#addbtn {
+  width: 25%;
+  border: 2px solid #000;
+  padding: 5px;
+  box-shadow: 2px 1px 8px #000;
+}
+
+.remove {
+  padding: 3px;
+  margin: 5px;
+  box-shadow: 2px 1px 8px #000;
+  border: 2px solid #000;
+}
+
+.card {
+  background-color: #fff;
+  display: flex;
+}
+
+.btn {
+  flex: 1;
+  text-align: end;
+}
+
+.p-content {
+  vertical-align: middle;
+}
+
+#message {
+  height: 10px;
+  display: none;
+  color: red;
+  font-size: small;
+  margin: 10px 0 10px;
+}
+
+.container {
+  display: block;
+}
+
+.contact-info-container {
+  display: none;
+}
+
+.add-book {
+  display: none;
+}
