+++ static/css/nav.css
@@ -1,115 +1,115 @@
 * {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-    font-family: "Indie Flower", cursive;
-    font-family: "Lato", sans-serif;
-    font-family: "Open Sans", sans-serif;
-    font-family: "Roboto", sans-serif;
-  }
-  
-  header {
-    display: flex;
-    border: 2px solid #000;
-    padding: 8px 20px 8px 20px;
-    margin: 12px 0 20px 0;
-  }
-  
-  .time {
-    text-align: right;
-  }
-  
-  .awesome-books {
-    flex: 1;
-  }
-  
-  .form-heading {
-    margin-bottom: 30px;
-  }
-  
-  .nav-links {
-    display: flex;
-    list-style-type: none;
-  }
-  
-  .link {
-    border-right: 1px solid #000;
-    padding: 5px 5px;
-    font-weight: bold;
-    font-family: "Indie Flower", cursive;
-    font-family: "Lato", sans-serif;
-    font-family: "Open Sans", sans-serif;
-    font-family: "Roboto", sans-serif;
-  }
-  
-  .link:hover {
-    color: rgb(4, 4, 243);
-    cursor: pointer;
-  }
-  
-  .link:focus {
-    color: rgb(4, 4, 243);
-  }
-  
-  .active {
-    display: block;
-  }
-  
-  .link:nth-child(3) {
-    border: none;
-  }
-  
-  .all-awesome-books {
-    text-align: center;
-    margin-bottom: 20px;
-    font-family: "Indie Flower", cursive;
-    font-family: "Lato", sans-serif;
-    font-family: "Open Sans", sans-serif;
-    font-family: "Roboto", sans-serif;
-  }
-  
-  .add-book,
-  .contact {
-    text-align: center;
-    display: block;
-  }
-  
-  .books-img {
-    text-align: center;
-    width: 40px;
-    height: 40px;
-  }
-  
-  .contact-info-container {
-    border: 3px solid #000;
-    padding: 15px;
-    text-align: start;
-    font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif;
-    line-height: 30px;
-  }
-  
-  .contact-info-heading {
-    font-size: 33px;
-    padding-bottom: 15px;
-    text-align: center;
-    margin: 20px 20px;
-  }
-  
-  footer {
-    margin-top: 3rem;
-  }
-  
-  .footer {
-    border: 3px solid #000;
-    padding: 20px;
-    width: 97%;
-    margin: 0 auto;
-  }
-  
-  .empty-footer {
-    width: 100%;
-    height: 50px;
-    background-color: rgb(163, 160, 160);
-    border: 3px solid #000;
-    margin-top: 15px;
-  }
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+  font-family: "Indie Flower", cursive;
+  font-family: "Lato", sans-serif;
+  font-family: "Open Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
+}
+
+header {
+  display: flex;
+  border: 2px solid #000;
+  padding: 8px 20px 8px 20px;
+  margin: 12px 0 20px 0;
+}
+
+.time {
+  text-align: right;
+}
+
+.awesome-books {
+  flex: 1;
+}
+
+.form-heading {
+  margin-bottom: 30px;
+}
+
+.nav-links {
+  display: flex;
+  list-style-type: none;
+}
+
+.link {
+  border-right: 1px solid #000;
+  padding: 5px 5px;
+  font-weight: bold;
+  font-family: "Indie Flower", cursive;
+  font-family: "Lato", sans-serif;
+  font-family: "Open Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
+}
+
+.link:hover {
+  color: rgb(4, 4, 243);
+  cursor: pointer;
+}
+
+.link:focus {
+  color: rgb(4, 4, 243);
+}
+
+.active {
+  display: block;
+}
+
+.link:nth-child(3) {
+  border: none;
+}
+
+.all-awesome-books {
+  text-align: center;
+  margin-bottom: 20px;
+  font-family: "Indie Flower", cursive;
+  font-family: "Lato", sans-serif;
+  font-family: "Open Sans", sans-serif;
+  font-family: "Roboto", sans-serif;
+}
+
+.add-book,
+.contact {
+  text-align: center;
+  display: block;
+}
+
+.books-img {
+  text-align: center;
+  width: 40px;
+  height: 40px;
+}
+
+.contact-info-container {
+  border: 3px solid #000;
+  padding: 15px;
+  text-align: start;
+  font-family: "Gill Sans", "Gill Sans MT", Calibri, "Trebuchet MS", sans-serif;
+  line-height: 30px;
+}
+
+.contact-info-heading {
+  font-size: 33px;
+  padding-bottom: 15px;
+  text-align: center;
+  margin: 20px 20px;
+}
+
+footer {
+  margin-top: 3rem;
+}
+
+.footer {
+  border: 3px solid #000;
+  padding: 20px;
+  width: 97%;
+  margin: 0 auto;
+}
+
+.empty-footer {
+  width: 100%;
+  height: 50px;
+  background-color: rgb(163, 160, 160);
+  border: 3px solid #000;
+  margin-top: 15px;
+}
