+++ static/js/nav.js
@@ -1,12 +1,9 @@
 /* eslint-disable no-unused-vars */
 /* eslint-disable global-require */
 
-
 const mainContainer = document.querySelector('.container');
 const addBookForm = document.querySelector('#form');
 const contactInfo = document.querySelector('.contact-info-container');
-const time = document.querySelector('.time');
-
 document.getElementById('add-book').addEventListener('click', () => {
   addBookForm.style.display = 'block';
   mainContainer.style.display = 'none';
@@ -25,21 +22,5 @@ document.getElementById('contactlink').addEventListener('click', () => {
   contactInfo.style.display = 'block';
 });
 
-const dateTable = ['janvier', 'fevrier', 'March', 'April', 'May', 'June', 'Jully', 'August', 'September', 'October', 'November', 'December'];
-
-function getTime() {
-  const globalDate = new Date();
-  const year = globalDate.getFullYear();
-  const hours = globalDate.getHours();
-  const minutes = globalDate.getMinutes();
-  const seconds = globalDate.getSeconds();
-  const day = globalDate.getDate();
-  let month = globalDate.getMonth();
-  month = dateTable[month];
-
-  time.innerHTML = `${month} ${day}th ${year}, ${hours}:${minutes}:${seconds}`;
-}
-
 window.addEventListener('load', () => {
-  setInterval(getTime, 1000);
 });
